+++ Spring Data/Spring Data Auto Mapping Objects/src/main/java/com/example/dtoExercise/model/dto/GameAddDto.java
@@ -8,7 +8,7 @@ import java.time.LocalDate;
 public class GameAddDto {
     private String title;
     private BigDecimal price;
-    private Integer size;
+    private Double size;
     private String trailer;
     private String thumbnailURL;
     private String description;
@@ -17,7 +17,7 @@ public class GameAddDto {
     public GameAddDto() {
     }
 
-    public GameAddDto(String title, BigDecimal price, Integer size, String trailer
+    public GameAddDto(String title, BigDecimal price, Double size, String trailer
             , String thumbnailURL, String description, String releaseDate) {
         this.title = title;
         this.price = price;
@@ -48,10 +48,10 @@ public class GameAddDto {
 
     //Validation positive price
     @Positive
-    public Integer getSize() {
+    public Double getSize() {
         return size;
     }
-    public void setSize(Integer size) {
+    public void setSize(Double size) {
         this.size = size;
     }
